pyproject.toml
[project]
name = "dc-pulse"
version = "0.1.0"
description = "Weekly Canadian DC pension trend digest pipeline"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "feedparser>=6.0.11",
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.6.0",
    "openai>=1.40.0",
    "python-dateutil>=2.9.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
